feat(use_cases/gif): add the number of rows to consider in the pagination

+++ backend/src/domain/entities/pagination.go
@@ -3,6 +3,6 @@ package entities
 type Pagination[T any] struct {
 	Page        int
 	RowsPerPage int
-	TotalRows   int
+	TotalRows   int64
 	Results     []*T
 }
